+++ nuttx/configs/mikroe-stm32f4/src/stm32_extmem.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/mikroe_stm32f4/src/up_extmem.c
- * arch/arm/src/board/up_extmem.c
+ * configs/mikroe_stm32f4/src/stm32_extmem.c
  *
  *   Copyright (C) 2012-2013 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
