+++ nuttx/configs/mikroe-stm32f4/src/stm32_spi.c
@@ -1,6 +1,5 @@
 /************************************************************************************
- * configs/mikroe_stm32f4/src/up_spi.c
- * arch/arm/src/board/up_spi.c
+ * configs/mikroe_stm32f4/src/stm32_spi.c
  *
  *   Copyright (C) 2011-2013 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -61,7 +60,7 @@
 #if defined(CONFIG_STM32_SPI1) || defined(CONFIG_STM32_SPI2) || defined(CONFIG_STM32_SPI3)
 
 /************************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ************************************************************************************/
 
 /* Enables debug output from this file (needs CONFIG_DEBUG too) */
